+++ CHANGES.md
@@ -1,5 +1,11 @@
 # zoql changelog
 
+## v1.0.2 (2018-06-25)
+
+ * Really correct `--describe`. Removed `explicitArray: false` from XML parsing
+   so I could have consistent, if annoying, parsed versions of the describe
+   response.
+
 ## v1.0.1 (2018-06-25)
 
  * Correct `--describe` when related objects is a single object.